added missing precondition and moved buffer.skip to new file
