+++ lib/peer.js
@@ -1,12 +1,13 @@
 'use strict';
 
-var Buffers = require('buffers');
+var Buffers = require('./buffers');
 var EventEmitter = require('events').EventEmitter;
 var Net = require('net');
 var Socks5Client = require('socks5-client');
 var bitcore = require('bitcore');
 var Networks = bitcore.Networks;
 var Messages = require('./messages');
+var $ = bitcore.util.preconditions;
 var util = require('util');
 
 /**
@@ -30,6 +31,9 @@ var util = require('util');
  * @constructor
  */
 function Peer(options) {
+  /* jshint maxstatements: 20 */
+  /* jshint maxcomplexity: 8 */
+
   if (!(this instanceof Peer)) {
     return new Peer(options);
   }
@@ -93,6 +97,8 @@ Peer.STATUS = {
  * @returns {Peer} The same Peer instance.
  */
 Peer.prototype.setProxy = function(host, port) {
+  $.checkState(this.status === Peer.STATUS.DISCONNECTED);
+
   this.proxy = {
     host: host,
     port: port
@@ -201,22 +207,6 @@ Peer.prototype._getSocket = function() {
   }
 
   return new Net.Socket();
-};
-
-// TODO: Remove this PATCH (yemel)
-Buffers.prototype.skip = function(i) {
-  if (i === 0) return;
-
-  if (i === this.length) {
-    this.buffers = [];
-    this.length = 0;
-    return;
-  }
-
-  var pos = this.pos(i);
-  this.buffers = this.buffers.slice(pos.buf);
-  this.buffers[0] = new Buffer(this.buffers[0].slice(pos.offset));
-  this.length -= i;
 };
 
 module.exports = Peer;
